requirements.txt
alembic==0.9.6
click==6.7
coverage==4.4.2
flake8==3.5.0
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.1
Werkzeug==0.14.1
WTForms==2.1
